refactor: move the preferences flags to new module
Move the preference flags over to the preferences module where they are used.
