+++ app/src/androidTest/java/com/chesire/nekome/features/ActivityTests.kt
@@ -2,8 +2,8 @@ package com.chesire.nekome.features
 
 import androidx.compose.ui.test.junit4.createComposeRule
 import com.chesire.nekome.UITest
-import com.chesire.nekome.core.flags.HomeScreenOptions
 import com.chesire.nekome.core.preferences.ApplicationPreferences
+import com.chesire.nekome.core.preferences.flags.HomeScreenOptions
 import com.chesire.nekome.robots.activity
 import com.chesire.nekome.robots.login.loginCredentials
 import com.chesire.nekome.robots.search.host
