+++ app/src/main/java/com/chesire/nekome/ActivityViewModel.kt
@@ -6,8 +6,8 @@ import androidx.lifecycle.asLiveData
 import androidx.lifecycle.viewModelScope
 import androidx.navigation.NavDirections
 import com.chesire.nekome.core.IOContext
-import com.chesire.nekome.core.flags.HomeScreenOptions
 import com.chesire.nekome.core.preferences.ApplicationPreferences
+import com.chesire.nekome.core.preferences.flags.HomeScreenOptions
 import com.chesire.nekome.datasource.auth.AccessTokenRepository
 import com.chesire.nekome.datasource.user.UserRepository
 import com.hadilq.liveevent.LiveEvent
