+++ core/preferences/src/test/java/com/chesire/nekome/core/preferences/ApplicationPreferencesTests.kt
@@ -3,9 +3,9 @@ package com.chesire.nekome.core.preferences
 import android.content.Context
 import android.content.SharedPreferences
 import com.chesire.nekome.core.R
-import com.chesire.nekome.core.flags.HomeScreenOptions
-import com.chesire.nekome.core.flags.Theme
 import com.chesire.nekome.core.flags.UserSeriesStatus
+import com.chesire.nekome.core.preferences.flags.HomeScreenOptions
+import com.chesire.nekome.core.preferences.flags.Theme
 import io.mockk.Runs
 import io.mockk.every
 import io.mockk.just
