+++ core/preferences/src/test/java/com/chesire/nekome/core/preferences/flags/HomeScreenOptionsTests.kt
@@ -1,4 +1,4 @@
-package com.chesire.nekome.core.flags
+package com.chesire.nekome.core.preferences.flags
 
 import android.content.Context
 import com.chesire.nekome.core.R
