+++ features/series/src/main/java/com/chesire/nekome/app/series/collection/core/CurrentSortUseCase.kt
@@ -1,7 +1,7 @@
 package com.chesire.nekome.app.series.collection.core
 
-import com.chesire.nekome.core.flags.SortOption
 import com.chesire.nekome.core.preferences.SeriesPreferences
+import com.chesire.nekome.core.preferences.flags.SortOption
 import javax.inject.Inject
 import kotlinx.coroutines.flow.first
