+++ features/settings/build.gradle
@@ -26,6 +26,7 @@ android {
 }
 
 dependencies {
+    implementation project(":core:preferences")
     implementation project(":core:resources")
     implementation project(":libraries:core")
